+++ ft_printf/pf_formatbackpad.c
@@ -6,7 +6,7 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/02/07 14:15:04 by pholster       #+#    #+#                */
-/*   Updated: 2019/04/24 14:15:07 by pholster      ########   odam.nl         */
+/*   Updated: 2019/07/20 20:12:37 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void		pf_formatbackpad(t_info *info)
 {
-	int		len;
-	int		i;
+	intmax_t	len;
+	intmax_t	i;
 
 	i = PF_PADADDED;
 	len = ft_max(0, PF_WIDTH) - PF_VAR_LEN;
