+++ ft_printf/pf_formatpad.c
@@ -6,13 +6,13 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/02/07 14:15:04 by pholster       #+#    #+#                */
-/*   Updated: 2019/04/24 14:15:22 by pholster      ########   odam.nl         */
+/*   Updated: 2019/07/20 20:13:33 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/ft_printf.h"
 
-static int	addprefix(t_info *info, char *str, size_t n)
+static size_t	addprefix(t_info *info, char *str, size_t n)
 {
 	if (n == 0)
 		return (0);
@@ -21,12 +21,12 @@ static int	addprefix(t_info *info, char *str, size_t n)
 		return (0);
 	if (PF_PRECISION > PF_WIDTH && ft_strequ(str, "0") == FALSE)
 		return (0);
-	return ((int)n);
+	return (n);
 }
 
-static void	addpad(t_info *info, int len, char c)
+static void		addpad(t_info *info, intmax_t len, char c)
 {
-	int		i;
+	intmax_t	i;
 
 	i = 0;
 	while (i < len)
@@ -36,9 +36,9 @@ static void	addpad(t_info *info, int len, char c)
 	}
 }
 
-static void	addzero(t_info *info, int len, int space, char *prfx)
+static void		addzero(t_info *info, intmax_t len, intmax_t space, char *prfx)
 {
-	int		prelen;
+	size_t	prelen;
 	char	c;
 
 	prelen = ft_strlen(prfx);
@@ -52,7 +52,7 @@ static void	addzero(t_info *info, int len, int space, char *prfx)
 		addpad(info, len, c);
 }
 
-static char	*getprefix(t_info *info)
+static char		*getprefix(t_info *info)
 {
 	if (PF_ISNEGATIVE)
 		return ("-");
@@ -68,12 +68,12 @@ static char	*getprefix(t_info *info)
 	return ("");
 }
 
-void		pf_formatpad(t_info *info)
+void			pf_formatpad(t_info *info)
 {
-	char	*prfx;
-	int		prelen;
-	int		len;
-	int		space;
+	char		*prfx;
+	size_t		prelen;
+	intmax_t	len;
+	intmax_t	space;
 
 	prfx = getprefix(info);
 	prelen = ft_strlen(prfx);
